requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.5.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==4.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.13
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
responses==0.18.0
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.13.3
torch==1.13.1+cu116
torchmetrics==1.7.3
torchvision==0.14.1+cu116
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.27.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.5.0
wandb==0.20.1
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1